pyproject.toml
[project]
name = "poetryvenvcreationapiv2"
version = "0.1.0"
description = "venv"
authors = [
    {name = "Mamy-cloud",email = "mamycompteprofessionnel@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi (>=0.135.0,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "pydantic[dotenv] (>=2.12.5,<3.0.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "starlette (>=0.52.1,<0.53.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
